Add Channel#ad_impressions (no options, no errors)
In principle, YouTube API should be able to return earnings and impressions
also for channels monetized through AdSense (without CMS).

In practice, this does not work, so the Yt gem will only run earnings and
impressions in the presence of a content owner.

Note that the encrypted environment variables are now too long to fit inside
.travis.yml and have been moved to travis by means of individually setting
them with `travis env set NAME value`.