Fix: always fetch next page and reset offset
Previously we would *stop* the iteration if the last page did not contain
enough items.

However, YouTube sometimes returns a `nextPageToken` and we have to check
the next page to ensure it's empty. So we will.

Additionally, every time we start looping again, we need to clear out the
old pageToken so we always start back from the first page.